Fix: _printf is Now Able to print Floating Point Numbers

+++ _identify_datatype.c
@@ -44,7 +44,7 @@ __attribute__((unused)) char *post_string)
 				end_replace_index += count;
 				return (datatype);
 			case 'f':              /* float */
-				datatype = "double";
+				datatype = "float";
 				end_replace_index += count;
 				return (datatype);
 			case 'i':              /* integer */
